Add crawl requests (#2)
### Changes
Adds a model for crawl requests, which will be used to queue up URLs to
be crawled.

### Why?

As of now, I'm thinking the user will eventually be able to submit a
URL to be crawled, and the system will queue up the URL to be crawled
by the crawler. The crawler will then fetch the page, parse it, and
store the response in S3. The metadata about the crawl response will be
stored in the database. But all this is still TBD.