Set up initial crawl behavior (#4)
### Changes

Adds initial crawl behavior and allows an admin to trigger it via a new
button.

### Why?

This PR just adds preliminary setup. We'll eventually need something
more advanced that can handle, e.g.

* Redirects and redirect chains
* Retry logic and throttling
* Storing request metadata, such as status codes, number of attempts
* Storing the body of a successful response in an external bucket

But for now these new additions allow us to focus on the happy path and
map out that behavior. We can address everything else as the need
arises.

### Screenshots


![Screenshot 2024-06-19 at 7 26
31 AM](https://github.com/seanmack/copydog-ai/assets/5815877/e7d8de94-d489-40ec-b820-fe760863ce9a)
![Screenshot 2024-06-19 at 7 26
53 AM](https://github.com/seanmack/copydog-ai/assets/5815877/266baa6f-1f44-43d8-a8b4-56a6e9b91e13)
![Screenshot 2024-06-19 at 7 27
36 AM](https://github.com/seanmack/copydog-ai/assets/5815877/b3e4fe30-3d1a-4d42-9794-2c6570ff01b9)