Remove unnecessary crawl_type column from crawl requests (#7)
### Changes

Removes `crawl_type` column from crawl requests.

### Why?

It seems like `crawl_type` won't be necessary, at least on the crawl
request model. We likely will have some abstraction that handles batch
crawls, but all that is still up in the air. Since this column is not
helping us now and might never, it's best to remove it.