Crawl job persists title tag (#8)
### Changes

When a crawl job is successful, the title tag is persisted. 


### Why?

We want to analyze and preserve key characteristics of the retrieved
HTML so we can use it later. For example, if users want to create a new
page that is similar to a crawled page, an analysis of the page will
help craft an apt prompt.

Upcoming changes will include other key attributes, such as headings,
content length, meta tags, etc.

Note: I considered performing this analysis on the fly so that we didn't
need to store these details in the database, but I think the overhead of
having to repeatedly analyze the HTML is worth avoiding. But if we
change our mind, the PageAnalysis class is flexible enough to be used
dynamically, so there's little risk here.

### Screenshots

![Screenshot 2024-06-20 at 7 38
23 AM](https://github.com/seanmack/copydog-ai/assets/5815877/ff691b53-897c-4c47-bafc-cea4e0cdee8d)