Use custom field for html_response (#10)
### Changes

Switches to conventional approach for displaying a custom field in an
Administrate dashboard.

### Why?

Previously, I customized the dashboard view so that it would provide a
download link for the `html_response` attachment. But Administrate has a
feature called "custom fields" and that approach is reusable, i.e. in
the future we can just use the new field type instead of manually adding
custom HTML.

### Screenshot

![Screenshot 2024-06-21 at 4 39
31 AM](https://github.com/seanmack/copydog-ai/assets/5815877/47830e84-1f62-4eea-b465-0dc48a4a7957)