Display sanitized html response (#12)
### Changes

User sees a sanitized version of the retrieved HTML for a given crawl
request.

### Why?

While we might want to extract details from the whole HTML response,
such as meta tags, we primarily want to work with the main content of
the page. The changes here add some preliminary sanitization processes
that allow us to dynamically extract only the content we care about.
We'll build on these processes soon to further distill the content.

### Notes

I used the `services/` directory for most of the new code. I'm not sure
that will make sense long term for all these classes, but while I'm
still shaping things, it works well enough.

### Screenshots

![Screenshot 2024-06-22 at 10 31
31 AM](https://github.com/seanmack/copydog-ai/assets/5815877/21d2ceda-14cb-4c6b-a320-536d40a451b0)