Persist api response (#15)
### Changes

The results of the API request to OpenAI are saved in a new model
`WebPageDraft`, which is associated to the crawl request.

### Why?

This is the next step toward the goal of providing a user with a clean
version of the primary content of a given webpage. Once we have that
locked in, we can tackle allowing the user to modify it, e.g. request
rewriting specific sections to suit their SEO goals.

### Screenshots

![Screenshot 2024-06-26 at 7 25
12 AM](https://github.com/seanmack/copydog-ai/assets/5815877/28a9df18-53d2-4cb9-8906-05e4adedadba)

![Screenshot 2024-06-26 at 7 27
24 AM](https://github.com/seanmack/copydog-ai/assets/5815877/e9680aa9-006b-48dd-8449-fc5a904fe52e)