Refactor crawl job logic (#17)
### Changes

Refactors the job that handles downloading and processing HTML content.

### Why?

The `handle_success` method was really long because I wanted to postpone
any refactors until after the entire flow was working as expected. Now
that everything is working well, it's time do some cleanup.

Note: I also added a minor tweak to improve a spec that was logging
warning messages when running the specs.