Create bulk crawl requests (#18)
### Changes

A user can add a list of URLs and create a new web page draft for each.
As with the individual page crawls, the fetched content is sanitized and
processed by the OpenAI API.

### Why?

Crawling URLs in bulk is a critical feature. It's typical for a website
owner to have a list of URLs they want to evaluate or clone, and this
new feature allows them to add a list of URLs and then generate a web
page draft for each.

### Notes

There are still some decisions that need to be made about how this
should work. Should users be able to trigger this more than once? If so,
should the associated CrawlRequests be updated or should new ones be
created?

Also, the admin view is getting cluttered. While it won't be
customer-facing, it does need to be functional, so it might be time soon
to start planning out where the various resources will live.

### Screenshots

![Screenshot 2024-06-29 at 5 53
45 PM](https://github.com/seanmack/copydog-ai/assets/5815877/8bfc0326-e4fa-42d0-9242-20c295d936ee)
![Screenshot 2024-06-29 at 5 53
59 PM](https://github.com/seanmack/copydog-ai/assets/5815877/955b81a5-0378-49f3-93d1-3ec289bfaf1e)
![Screenshot 2024-06-29 at 5 54
58 PM](https://github.com/seanmack/copydog-ai/assets/5815877/310b2d66-3c52-41c4-ac67-066a545ee819)
![Screenshot 2024-06-29 at 5 56
10 PM](https://github.com/seanmack/copydog-ai/assets/5815877/71a57a00-91b9-4c46-b673-def15b60709f)