Add script to condense app code
This script condenses the code of this app into a single text file. It
excludes anything sensitive, like credentials, as well as anything
not very useful.

The goal is to be able to share the code of this app with ChatGPT or
Claude to get feedback or to get help with a specific
problem. I'm still testing this out, so I'm probably going to make
continual changes to this. And if it works well, I probably will move
this script out of the app, maybe into a gem or something.